Room price filtering

+++ Craciun_Adriana_HotelReservation/Models/Reservation.cs
@@ -7,7 +7,7 @@
         public int RoomId { get; set; }
         public DateTime CheckInDate { get; set; }
         public DateTime CheckOutDate { get; set; }
-        public decimal TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public ReservationStatus Status { get; set; }
         public string? SpecialRequests { get; set; }
         public MarketSegmentType MarketSegment { get; set; }
