+++ Craciun_Adriana_HotelReservation/Models/Room.cs
@@ -7,7 +7,7 @@
         public string RoomNumber { get; set; }
         public RoomType RoomType { get; set; } // Single, Double, Suite
         public int Capacity { get; set; }
-        public decimal BasePricePerNight { get; set; }
+        public float BasePricePerNight { get; set; }
         public Hotel? Hotel { get; set; }
         public ICollection<Reservation>? Reservations { get; set; }
